Refactor tests: Move tests from module and typed array files to dedicated test modules
- Moved `into_js_module` and `can_throw_exception` tests from `core/engine/src/module/mod.rs` to `core/engine/src/module/tests/mod.rs`.
- Moved typed array iterator tests from `core/engine/src/object/builtins/jstypedarray.rs` to `core/engine/src/object/builtins/jstypedarray/tests.rs`.
- Removed unused test module from `core/engine/src/value/conversions/serde_json.rs`.
- Moved NanBoxedValue tests from `core/engine/src/value/inner/nan_boxed.rs` to `core/engine/src/value/inner/nan_boxed/tests.rs`.
