+++ core/engine/src/value/conversions/serde_json.rs
@@ -204,5 +204,3 @@ impl JsValue {
     }
 }
 
-#[cfg(test)]
-mod tests;
